:recycle: refactor(BTS-300): http 폴더 위치 변경 shared/api
