+++ src/features/study-rooms/components/sidebar/services/api.ts
@@ -1,5 +1,5 @@
 // 수업노트 그룹 생성
-import { authApi } from '@/shared/lib/http/http.client';
+import { authApi } from '@/shared/api';
 
 export const createStudyNoteGroup = async (params: {
   studyRoomId: number;
