+++ src/shared/lib/index.ts
@@ -1,5 +1,4 @@
 export * from './bff';
-export * from './http';
 export * from './server';
 export * from './error';
 export * from './message';
